# coremodels: Change input filename, optimize imports