… pkg/coremodel/dashboard/coremodel.cue