<!--

Thank you for sending a pull request! Here are some tips:

1. If this is your first time, please read our contribution guide at https://github.com/grafana/grafana/blob/main/CONTRIBUTING.md

2. Ensure you include and run the appropriate tests as part of your Pull Request.

3. In a new feature or configuration option, an update to the documentation is necessary. Everything related to the documentation is under the docs folder in the root of the repository.

4. If the Pull Request is a work in progress, make use of GitHub's "Draft PR" feature and mark it as such.

5. If you can not merge your Pull Request due to a merge conflict, Rebase it. This gets it in sync with the main branch.

6. Name your PR as "<FeatureArea>: Describe your change", e.g. Alerting: Prevent race condition. If it's a fix or feature relevant for the changelog describe the user impact in the title. The PR title is used to auto-generate the changelog for issues marked with the "add to changelog" label.

-->

**What this PR does / why we need it**:

Two small things:

First, this changes the filename in which a coremodel is declared from `pkg/coremodel/<name>/lineage.cue` to `pkg/coremodel/<name>/coremodel.cue`. In working on the design doc explaining this whole system, it's become clear that it's clearer if we refer to that declaration as "the coremodel."

Second, this removes a bunch of the pointless import statements that were being generated, then removed by `goimports`. This speeds up coremodel codegen on my laptop by a downright comical amount:
```
# Before
go generate  13.93s user 31.39s system 441% cpu 10.265 total
# After
go generate  1.62s user 1.07s system 89% cpu 2.990 total
```